requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.5
django-environ==0.4.5
docutils==0.16
factory-boy==3.1.0
Faker==4.0.3
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.4.4
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.9
zipp==3.1.0
